Extracting index config files into separate folder
